+++ dll/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "lepton_jpeg_0_5"
+version = "0.5.0"
+edition = "2021"
+authors = ["Kristof Roomp <kristofr@microsoft.com>"]
+
+[dependencies]
+lepton_jpeg = { path = "../lib" }
+rstest = "0.22"
+
+[lib]
+crate-type = ["cdylib"]
+
+
