+++ lib/src/consts.rs
