+++ lib/src/jpeg/bit_reader.rs
