+++ lib/src/jpeg/bit_writer.rs
