+++ lib/src/jpeg/block_based_image.rs
