+++ lib/src/jpeg/component_info.rs
