+++ lib/src/jpeg/jpeg_code.rs
