+++ lib/src/jpeg/jpeg_position_state.rs
