+++ lib/src/jpeg/row_spec.rs
