+++ lib/src/jpeg/truncate_components.rs
