+++ lib/src/structs/lepton_encoder.rs
