+++ lib/src/structs/lepton_file_writer.rs
