+++ lib/src/structs/model.rs
