+++ lib/src/structs/partial_buffer.rs
