+++ lib/src/structs/probability_tables.rs
