+++ lib/src/structs/simple_hash.rs
