+++ lib/src/structs/vpx_bool_writer.rs
